requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
celery>=5.3.0
click>=8.1.0
aiohttp>=3.8.0
redis>=4.5.0
channels>=4.0.0
daphne>=4.0.0
channels-redis>=4.1.0
dashscope>=1.18.0
python-dotenv>=1.0.0
pandas>=2.0.0
openpyxl>=3.0.0
xlrd>=2.0.0
Pillow>=9.0.0
langchain>=0.3.0
langchain-core>=0.3.67     # 必须 >=0.3.67 以兼容 langgraph/openai
langchain-community>=0.3.0
langchain-openai>=0.1.0   # 如果你用 OpenAI 兼容模式（推荐！）
